Create image processing project1
Some simple image processing operations were done in this project. The user must upload the image to be processed. This study overlaps the operations selected by the user. At some points, there are processes and incomplete operating situations that differ from this situation. These situations are explained in the readme tab.